design: 배경색상을 적용한다
