+++ src/components/quick-link/QuickLink.tsx
@@ -1,34 +1,9 @@
-import React from "react";
-import SvgIcon, { SvgIconProps } from "@mui/material/SvgIcon";
 import styled from "@emotion/styled";
 import LibraryBooksRoundedIcon from "@mui/icons-material/LibraryBooksRounded";
 import AccountBoxRoundedIcon from "@mui/icons-material/AccountBoxRounded";
 import SmsRoundedIcon from "@mui/icons-material/SmsRounded";
 import CreateRoundedIcon from "@mui/icons-material/CreateRounded";
-
-const QuickLinkIcon = styled(SvgIcon)<SvgIconProps>`
-  color: var(--gray5-lowText);
-  width: 25px;
-  height: 25px;
-`;
-
-const QuickLinkText = styled.span`
-  color: var(--gray5-lowText);
-  display: inline-block;
-  text-align: center;
-`;
-
-const QuickLinkIconTextItem = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 1rem;
-  width: 279px;
-  height: 80px;
-  border-right: 1px solid var(--gray1-background);
-  box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
-  flex: 1;
-`;
+import QuickLinkItem from "./QuickLinkItem";
 
 const QuickLinkWrapper = styled.div`
   display: flex;
@@ -36,19 +11,9 @@ const QuickLinkWrapper = styled.div`
   width: 90%;
   max-width: 1116px;
   box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+  background-color: var(--gray1-background);
+  cursor: pointer;
 `;
-
-interface QuickLinkItemProps {
-  icon: React.ComponentType<SvgIconProps>;
-  text: string;
-}
-
-const QuickLinkItem = ({ icon: IconComponent, text }: QuickLinkItemProps) => (
-  <QuickLinkIconTextItem>
-    <QuickLinkIcon component={IconComponent} />
-    <QuickLinkText>{text}</QuickLinkText>
-  </QuickLinkIconTextItem>
-);
 
 const QuickLink = () => (
   <QuickLinkWrapper>
